+++ lib/src/codecs/byte/byte_encoder.dart
@@ -1,0 +1,11 @@
+import 'dart:convert';
+
+class XsdByteEncoder extends Converter<int, String> {
+  const XsdByteEncoder();
+
+  @override
+  String convert(int input) {
+    if (input == 0) return "0";
+    return input.toString();
+  }
+}