+++ lib/src/codecs/codecs.dart
@@ -7,3 +7,4 @@ export 'token/token_codec.dart';
 export 'int/int_codec.dart';
 export 'ncname/ncname_codec.dart';
 export 'short/short_codec.dart';
+export 'byte/byte_codec.dart';